+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.3.x
+
+## Bug fixes
+- PIM-4082: Fix error translation keys when creating a job
+
 # 1.3.8 (2015-04-14)
 
 ## Bug fixes